requirements.txt
tensorflow==2.5.0
h5py==3.1.0
imgaug==0.4.0
keras==2.9.0
manga_ocr==0.1.8
matplotlib==3.2.2
numpy==1.19.5
opencv_python_headless==4.6.0.66
Pillow==9.3.0
requests==2.23.0
scikit_image==0.18.3
scipy==1.7.3
streamlit==1.13.0
urllib3==1.24.3
